Revert "Bulk Promo Codes"
