--- api/app/serializers/spree/api/v2/platform/promotion_batch_serializer.rb
@@ -1,15 +1,0 @@
-module Spree
-  module Api
-    module V2
-      module Platform
-        class PromotionBatchSerializer < BaseSerializer
-          include ResourceSerializerConcern
-
-          belongs_to :template_promotion, serializer: PromotionSerializer
-
-          has_many :promotions
-        end
-      end
-    end
-  end
-end
