--- core/app/services/spree/promotion_batches/generate_code.rb
@@ -1,17 +1,0 @@
-module Spree
-  module PromotionBatches
-    class GenerateCode
-      def initialize(random: SecureRandom)
-        @random = random
-      end
-
-      def call(random_characters:, prefix: nil, suffix: nil)
-        [
-          prefix,
-          @random.hex(random_characters),
-          suffix
-        ].join('')
-      end
-    end
-  end
-end
