--- core/db/migrate/20231114182646_add_promotion_batch_reference_to_promotions.rb
@@ -1,6 +1,0 @@
-class AddPromotionBatchReferenceToPromotions < ActiveRecord::Migration[6.1]
-  def change
-    add_reference :spree_promotions, :promotion_batch, foreign_key: { to_table: :spree_promotion_batches}
-    add_column :spree_promotions, :template, :boolean
-  end
-end
