+++ core/lib/spree/core/dependencies.rb
@@ -97,10 +97,7 @@ module Spree
         credit_card_finder: 'Spree::CreditCards::Find',
         products_finder: 'Spree::Products::Find',
         taxon_finder: 'Spree::Taxons::Find',
-        line_item_by_variant_finder: 'Spree::LineItems::FindByVariant',
-
-        # promotion batches
-        promotion_batch_destroy_service: 'Spree::PromotionBatches::Destroy'
+        line_item_by_variant_finder: 'Spree::LineItems::FindByVariant'
       }.freeze
 
       include Spree::DependenciesHelper
